requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
joblib==1.3.2
nltk==3.8.1
numpy==1.25.2
pandas==2.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
regex==2023.8.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
